refactor: Moved rdf_to_graph
