+++ cognite/neat/graph/extractors/__init__.py
@@ -1,4 +1,3 @@
-from cognite.neat.graph.extractors import rdf_to_graph
 from cognite.neat.graph.extractors._graph_capturing_sheet import (
     extract_graph_from_sheet,
     read_graph_excel_file_to_table_by_name,
@@ -9,7 +8,6 @@ from ._graph_capturing_sheet import GraphCapturingSheet
 
 __all__ = [
     "MockGraphGenerator",
-    "rdf_to_graph",
     "extract_graph_from_sheet",
     "read_graph_excel_file_to_table_by_name",
     "GraphCapturingSheet",
