+++ cognite/neat/graph/stores/__init__.py
@@ -1,4 +1,5 @@
 from .configuration import RdfStoreConfig, RdfStoreType
 from .graph_store import NeatGraphStore, drop_graph_store_storage
+from .rdf_to_graph import rdf_file_to_graph
 
-__all__ = ["NeatGraphStore", "drop_graph_store_storage", "RdfStoreType", "RdfStoreConfig"]
+__all__ = ["NeatGraphStore", "drop_graph_store_storage", "RdfStoreType", "RdfStoreConfig", "rdf_file_to_graph"]
